feat: blacktea() factory and pay() function (T1)
The orchestrator. Wires schema + evaluator + idempotency + history + rails
into one PayFunction the agent calls.

- src/agent.ts: blacktea(opts) returns pay(input, opts?). Internally:
  - validates the input
  - checks idempotency cache (returns cached receipt on hit)
  - picks a rail (first supports() match, or opts.rail override)
  - evaluates the policy
  - branches on decision: allow / approval / reject
  - approval channels: console (CLI prompt) or callback (onApprovalNeeded)
  - dry-run skips the rail and returns a simulated receipt
  - records to history on success
  - emits audit events at every step
- src/policy/load.ts: loadPolicy() accepts a Policy object or a file path.
- src/index.ts: public exports (factory, types, errors, schema, defaults).
- decision.ts evaluator.ts updated: Decision now carries `rule_fired`
  (rule[N] or "default") so audit logs and ApprovalRequest can reference
  the matching rule by position.

25 new agent tests cover happy path, reject, approval (approve / deny /
timeout / missing-handler), idempotency, dry-run, rail selection, and
input validation. 109 tests total. Lint, typecheck, build all clean.

+++ src/policy/schema.ts
@@ -83,7 +83,7 @@ export const ActionSchema = z.union([
         z
           .object({
             via: ApprovalChannelSchema,
-            timeout_seconds: z.number().int().positive().optional(),
+            timeout_seconds: z.number().positive().optional(),
           })
           .strict(),
       ]),
